advance

+++ burn.php
@@ -13,22 +13,27 @@ $mid = '';
 $action=$_GET['action'];
 if($action == 'get'){
 	$mid = $_GET['mid'];
+	$vc = $_GET['vc'];
 	if(strlen($mid)<30)
-		die('{"response":"failure","type":"SERVER PAGE ERROR : bad request"}');	
+		die('{"response":"failure","type":"SERVER PAGE ERROR : bad request"}');
+	if(strlen($vc) != 4)
+		die('{"response":"failure","type":"SERVER PAGE ERROR : bad request"}');		
 }
 
 
 if ($dir = opendir($data_dir)) {
     while (false !== ($file = readdir($dir))) {
         if ($file != "." && $file != ".." && $file != "index.html") {
-        		list($fmid, $fdate, $fexipired, $ftype) = explode('.', $file);
+        		list($fvc,$fmid, $fdate, $fexipired, $ftype) = explode('.', $file);
         		if($fdate+$fexipired*60 < $now){
         			unlink($data_dir.$file);
         			//die("expired".$data_dir.$file);
         			$fmid = 'x';
+        			$fvc = 'x';
         			$continue;
         		}
-        		if($mid == $fmid ){
+        		//echo $fvc."   ".$fmid."<br/>";
+        		if($mid == $fmid && $vc==$fvc){
         			break;	
         		}
             $count++;
@@ -36,12 +41,12 @@ if ($dir = opendir($data_dir)) {
     }
     closedir($dir);
 }
-if($action == 'get' && $mid == $fmid){
+if($action == 'get' && $mid == $fmid && $vc==$fvc){
 	$remain = $fdate+$fexipired*60 - $now;
 	echo('{"response":"success","type":"'.$ftype.'","remain":"'.$remain.'","data":"'.file_get_contents($data_dir.$file).'"}');	
 	unlink($data_dir.$file);
 }
-if($action == 'get' && $mid != $fmid){
+if($action == 'get' && ($mid != $fmid || $vc!=$fvc)){
 	die('{"response":"failure","message":"SERVER PAGE ERROR : no data"}');	
 }
